+++ Database.hpp
@@ -6,8 +6,8 @@ namespace locke {
 
 class Database {
  public:
-  int static get(void*, char*, uint32_t, int);
-  int static set(void*, char*, uint32_t, int);
+  int static get(void*, char*, uint32_t, size_t);
+  int static set(void*, char*, uint32_t, size_t);
 };
 
 }
